Move http fetch request to custom hook to avoid code repitition
