+++ src/containers/Standings.js
@@ -1,8 +1,9 @@
 import { useDispatch, useSelector } from 'react-redux';
-import { useEffect } from 'react';
+// import { useEffect } from 'react';
 import { Link } from 'react-router-dom';
+import useHttp from '../hooks/http';
 import TableFilter from '../components/TableFilter';
-import { refreshTable, filterTable } from '../actions';
+import { filterTable } from '../actions';
 import getStandingsByFilter from '../logic/filterFuncs';
 
 const Standings = () => {
@@ -12,25 +13,7 @@ const Standings = () => {
   const filter = useSelector((state) => state.filter);
   const dispatch = useDispatch();
 
-  useEffect(() => {
-    fetch('https://api-football-v1.p.rapidapi.com/v3/standings?season=2021&league=140', {
-      method: 'GET',
-      headers: {
-        'x-rapidapi-host': 'api-football-v1.p.rapidapi.com',
-        'x-rapidapi-key': 'ac39b316e0msh7d2b82002789670p10a6e9jsna343b2d68ca3',
-      },
-      mode: 'cors',
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        dispatch(refreshTable(data.response));
-      })
-      .catch((err) => {
-        console.error(err);
-      });
-
-    console.log('Standings Effects');
-  }, []);
+  useHttp('https://api-football-v1.p.rapidapi.com/v3/standings?season=2021&league=140', []);
 
   const handleFilterChange = (filter) => {
     dispatch(filterTable(filter));
